+++ src/main/java/net/burningtnt/webp/vp8l/transform/PredictorTransform.java
@@ -73,18 +73,16 @@ public final class PredictorTransform implements Transform {
 
     @Override
     public void applyInverse(RGBABuffer raster) {
-        int width = (int) raster.getWidth();
-        int height = (int) raster.getHeight();
+        int width = raster.getWidth();
+        int height = raster.getHeight();
 
         byte[] rgba = new byte[4];
 
         // Handle top and left border separately
 
         // (0,0) Black (0x000000ff) predict
-        // raster.getDataElements(0, 0, rgba);
         raster.getDataElements(0, 0, rgba);
         rgba[3] += 0xff;
-        // raster.setDataElements(0, 0, rgba);
         raster.setDataElements(0, 0, rgba);
 
         byte[] predictor = new byte[4];
@@ -93,34 +91,23 @@ public final class PredictorTransform implements Transform {
 
         // (x,0) L predict
         for (int x = 1; x < width; x++) {
-//            raster.getDataElements(x, 0, rgba);
             raster.getDataElements(x, 0, rgba);
-            //            raster.getDataElements(x - 1, 0, predictor);
             raster.getDataElements(x - 1, 0, predictor);
             addPixels(rgba, predictor);
-
-//            raster.setDataElements(x, 0, rgba);
             raster.setDataElements(x, 0, rgba);
         }
 
         // (0,y) T predict
         for (int y = 1; y < height; y++) {
-//            raster.getDataElements(0, y, rgba);
             raster.getDataElements(0, y, rgba);
-            //            raster.getDataElements(0, y - 1, predictor);
             raster.getDataElements(0, y - 1, predictor);
             addPixels(rgba, predictor);
-
-//            raster.setDataElements(0, y, rgba);
             raster.setDataElements(0, y, rgba);
         }
 
         for (int y = 1; y < height; y++) {
             for (int x = 1; x < width; x++) {
-//                int transformType = data.getSample(x >> bits, y >> bits, 1);
                 int transformType = data.getSample(x >> bits, y >> bits, 1);
-
-//                raster.getDataElements(x, y, rgba);
                 raster.getDataElements(x, y, rgba);
 
                 int lX = x - 1; // x for left
@@ -135,84 +122,63 @@ public final class PredictorTransform implements Transform {
                         rgba[3] += 0xff;
                         break;
                     case PREDICTOR_L:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_T:
-//                        raster.getDataElements(x, tY, predictor);
                         raster.getDataElements(x, tY, predictor);
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_TR:
-//                        raster.getDataElements(trX, trY, predictor);
                         raster.getDataElements(trX, trY, predictor);
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_TL:
-//                        raster.getDataElements(lX, tY, predictor);
                         raster.getDataElements(lX, tY, predictor);
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_AVG_L_TR_T:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(trX, trY, predictor2);
                         raster.getDataElements(trX, trY, predictor2);
                         average2(predictor, predictor2);
-
-//                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
                         average2(predictor, predictor2);
 
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_AVG_L_TL:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(lX, tY, predictor2);
                         raster.getDataElements(lX, tY, predictor2);
                         average2(predictor, predictor2);
 
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_AVG_L_T:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
                         average2(predictor, predictor2);
 
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_AVG_TL_T:
-//                        raster.getDataElements(lX, tY, predictor);
                         raster.getDataElements(lX, tY, predictor);
-                        //                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
                         average2(predictor, predictor2);
 
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_AVG_T_TR:
-//                        raster.getDataElements(x, tY, predictor);
                         raster.getDataElements(x, tY, predictor);
-                        //                        raster.getDataElements(trX, trY, predictor2);
                         raster.getDataElements(trX, trY, predictor2);
                         average2(predictor, predictor2);
 
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_AVG_L_TL_T_TR:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(lX, tY, predictor2);
                         raster.getDataElements(lX, tY, predictor2);
                         average2(predictor, predictor2);
-
-//                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
-                        //                        raster.getDataElements(trX, trY, predictor3);
                         raster.getDataElements(trX, trY, predictor3);
                         average2(predictor2, predictor3);
 
@@ -221,35 +187,25 @@ public final class PredictorTransform implements Transform {
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_SELECT:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
-                        //                        raster.getDataElements(lX, tY, predictor3);
                         raster.getDataElements(lX, tY, predictor3);
 
 
                         addPixels(rgba, select(predictor, predictor2, predictor3));
                         break;
                     case PREDICTOR_CLAMP_ADD_SUB_FULL:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
-                        //                        raster.getDataElements(lX, tY, predictor3);
                         raster.getDataElements(lX, tY, predictor3);
                         clampAddSubtractFull(predictor, predictor2, predictor3);
 
                         addPixels(rgba, predictor);
                         break;
                     case PREDICTOR_CLAMP_ADD_SUB_HALF:
-//                        raster.getDataElements(lX, y, predictor);
                         raster.getDataElements(lX, y, predictor);
-                        //                        raster.getDataElements(x, tY, predictor2);
                         raster.getDataElements(x, tY, predictor2);
                         average2(predictor, predictor2);
-
-//                        raster.getDataElements(lX, tY, predictor2);
                         raster.getDataElements(lX, tY, predictor2);
                         clampAddSubtractHalf(predictor, predictor2);
 
@@ -257,8 +213,6 @@ public final class PredictorTransform implements Transform {
                         break;
 
                 }
-
-//                raster.setDataElements(x, y, rgba);
                 raster.setDataElements(x, y, rgba);
             }
         }
